// web/src/openapi/models/TracingLevel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tracing-lv-server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TracingLevel = {
    Trace: 'Trace',
    Debug: 'Debug',
    Info: 'Info',
    Warn: 'Warn',
    Error: 'Error'
} as const;
export type TracingLevel = typeof TracingLevel[keyof typeof TracingLevel];


export function instanceOfTracingLevel(value: any): boolean {
    for (const key in TracingLevel) {
        if (Object.prototype.hasOwnProperty.call(TracingLevel, key)) {
            if (TracingLevel[key as keyof typeof TracingLevel] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TracingLevelFromJSON(json: any): TracingLevel {
    return TracingLevelFromJSONTyped(json, false);
}

export function TracingLevelFromJSONTyped(json: any, ignoreDiscriminator: boolean): TracingLevel {
    return json as TracingLevel;
}

export function TracingLevelToJSON(value?: TracingLevel | null): any {
    return value as any;
}

export function TracingLevelToJSONTyped(value: any, ignoreDiscriminator: boolean): TracingLevel {
    return value as TracingLevel;
}